rag_api/requirements.txt
google-generativeai==0.8.5
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
chromadb==0.5.5
pypdf==5.1.0
python-docx==1.1.2
